refactor(InputValidator): extract inputView validation logic to new class
