+++ src/main/java/racingcar/view/InputView.java
@@ -7,20 +7,12 @@ import java.util.stream.Collectors;
 
 public class InputView {
 	private static final String DELIMITER = ",";
-	private static final int ROUND_MIN = 1;
-	private static final int ROUND_MAX = 10;
 	private static final Scanner scanner = new Scanner(System.in);
 
 	public static List<String> readCarNames() {
 		final String carNames = scanner.nextLine();
-		validateCarNames(carNames);
+		InputValidator.validateBlankedNames(carNames);
 		return splitByDelimAndTrimNames(carNames, DELIMITER);
-	}
-
-	private static void validateCarNames(final String input) {
-		if (input.isBlank()) {
-			throw new IllegalArgumentException("[ERROR]: 공백을 입력할 수 없습니다.");
-		}
 	}
 
 	private static List<String> splitByDelimAndTrimNames(final String names, final String delimiter) {
@@ -31,31 +23,8 @@ public class InputView {
 
 	public static int readRoundCount() {
 		final String roundCount = scanner.nextLine();
-		validateRoundCount(roundCount);
+		InputValidator.validateRoundCount(roundCount);
 		System.out.println();
 		return Integer.parseInt(roundCount);
-	}
-
-	private static void validateRoundCount(final String input) {
-		if (!isInputInGivenRange(input, ROUND_MIN, ROUND_MAX)) {
-			throw new IllegalArgumentException("[ERROR]: 1 이상 10 이하의 자연수가 입력되어야 합니다.");
-		}
-	}
-
-	private static boolean isInputInGivenRange(final String input, final int minInclusive, final int maxInclusive) {
-		int parsedInput;
-		try {
-			parsedInput = Integer.parseInt(input);
-			return isNumberInGivenRange(parsedInput, minInclusive, maxInclusive);
-		} catch (IllegalArgumentException e) {
-		}
-		return false;
-	}
-
-	private static boolean isNumberInGivenRange(final int num, final int minInclusive, final int maxInclusive) {
-		if (num >= minInclusive && num <= maxInclusive) {
-			return true;
-		}
-		return false;
 	}
 }
